change dev package name to cmake package name
as discussed at munich sprint.

rationale:
The majority of KF5 libraries are one library with a matching cmake package
name (e.g. kconfig = libkf5config = cmake/kf5config). To a user this
paints a picture where this equality is always true so that if
cmake complains about missing KF5Wayland they want to install
libkf5wayland-dev to meet that requirement. This will however fail because
of the general naming rule we employ (i.e. single lib => libname-dev;
multilib => sourcename-dev) as kwayland produces two libraries its dev
package is called kwayland-dev rather than libkf5wayland-dev. This is
not obvious to users and causes to them unreasonable inconsistency.
Additionally it increases the entry barrier to packaging as a whole as
one needs to be aware of the -dev naming rule or else multilib sources
will seem equally out of line with the rest of frameworks. To that extend
even knowing this a packager would have to remember the frameworks that
are multilib or have to look up the correct names each time when writing
a new control file or expanding it with new build depends.
To increase overall consistency all frameworks are now to reflect their
cmake package name in the -dev package name. This is because cmake is
expected to be the primary method of consumption for frameworks and even
when using qmake the metadata is split per-library anyway, so they have
no common naming consistency underlying them.

This change updates the -dev name to align with expectation and introduces
a transitional dummy package to help with the transition.